Delete origin response `content-length` header
We're re-encoding the body and the Node.js server sends
the body however it wants. We should not interfere with
that.

This fixes and issue where sometimes the response is
cut of because we still forward the `content-length`
header even though Node.js is sending the body
with transfer-encoding: chunked.