Extract viewer data separately so we can retain more details
Until now we only constructed a viewer request object and we
were limited by the format offerred for that. There was no
way to retain the original HTTP version or protocol.

This resulted in a big where we could not accurately determine
the request protocol when setting the `CloudFront-Viewer-*` headers
and we hard-coded it to `http`. The same applies to the HTTP
version.

With this change, viewer data is extracted separately into
a separate object and passed when we construct the origin request.
Like this we aren't limited by structure of the `CloudFrontRequest`
interface.

This also fixes the bugs with the HTTP protocol and HTTP version.